ipq806x: support rpm clocks scaling
Add the driver that scales APPS and DDR fabric clocks according to
CPU frequency.

This is a reworked/simplified version of QSDK fab-scaling driver [1]
that utilizes upstream kernel qcom rpm-clk driver..

I've decided to keep it as a separate driver not to pollute the
CPUFREQ-DT code.
Scaling happens on the following logic:
APPS and DDR nominal frequency is 400mhz, max frequency is 533mhz.
Whenever CPU scales to > 1ghz then those fabrics are scaled to max
frequency. Those values are adjustable through the DT.

[1] https://source.codeaurora.org/quic/qsdk/oss/kernel/linux-msm/commit/?h=eggplant&id=7f4d9b5c8814329a66fe44d0dac55a4bd3cbcb78

Signed-off-by: Pavel Kubelun <be.dissent@gmail.com>